update speedex to include blocks with validation failures in its VM chain;  blocks with validation failures are included in the header chain, and validation status stored in BlockHeaderHashMap.  Bad headers are partially corrected -- rehashed AND blockNumber is made consistent.  Motivation is to enable recovery (relying on hotstuff data) from a shorn read (e.g. memorydatabase persisted but orderbooks did not -- can't distinguish validation failures just from hotstuff log in this case

+++ speedex/speedex_options.h
@@ -1,14 +1,18 @@
 #pragma once
+
 #include <cstddef>
+#include <cstdint>
+
+#include "speedex/approximation_parameters.h"
 
 namespace speedex {
 
 struct SpeedexOptions {
 
 	// protocol parameters
-	unsigned int tax_rate;
-	unsigned int smooth_mult;
-	unsigned int num_assets;
+	uint8_t tax_rate;
+	uint8_t smooth_mult;
+	uint16_t num_assets;
 
 	// operational parameters
 	size_t persistence_frequency;
